+++ src/lib.rs
@@ -1,5 +1,6 @@
 //! CL4SE application library.
 
 pub mod config;
+pub mod control;
 pub mod core;
 pub mod platform;